+++ SensateIoT.Platform.Router.Common/Init/RouterInitExtensions.cs
@@ -6,13 +6,17 @@
  */
 
 using System;
+
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+
 using SensateIoT.Platform.Router.Common.Caching.Abstract;
 using SensateIoT.Platform.Router.Common.Collections.Abstract;
 using SensateIoT.Platform.Router.Common.Routing;
 using SensateIoT.Platform.Router.Common.Routing.Abstract;
 using SensateIoT.Platform.Router.Common.Routing.Routers;
+using SensateIoT.Platform.Router.Common.Settings;
 using SensateIoT.Platform.Router.Data.Abstract;
 
 namespace SensateIoT.Platform.Router.Common.Init
@@ -33,7 +37,8 @@ namespace SensateIoT.Platform.Router.Common.Init
 				var queue = provider.GetRequiredService<IRemoteNetworkEventQueue>();
 				var inputQueue = provider.GetRequiredService<IQueue<IPlatformMessage>>();
 				var logger = provider.GetRequiredService<ILogger<CompositeRouter>>();
-				var router = new CompositeRouter(cache, inputQueue, queue, logger) as IMessageRouter;
+				var options = provider.GetRequiredService<IOptions<RoutingQueueSettings>>();
+				var router = new CompositeRouter(cache, inputQueue, queue, options, logger) as IMessageRouter;
 
 				AddRouters(router, provider);