+++ SensateIoT.Platform.Router.Common/Routing/Abstract/IMessageRouter.cs
@@ -6,12 +6,13 @@
  */
 
 using System;
+using System.Threading;
 
 namespace SensateIoT.Platform.Router.Common.Routing.Abstract
 {
 	public interface IMessageRouter : IDisposable
 	{
 		void AddRouter(IRouter router);
-		bool TryRoute();
+		bool TryRoute(CancellationToken token = default);
 	}
 }